requirements.txt
python-docx>=0.8.11
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.2
gunicorn>=21.0.0
